third_party/WebKit/LayoutTests/overflow/overflow-transform-001.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#container {
  overflow: auto;
  border: 1px solid black;
  width: 200px;
  height: 100px;
}
#target {
  width: 300px;
  height: 200px;
  background-color: rgba(0, 255, 0, 0.3);
  transform: translateX(7px) translateY(9px)
}
</style>
<p>Overflow, transform 2D</p>
<div id="container">
  <div id="target"></div>
</div>
<script>

var container = document.querySelector('#container');
var target = document.querySelector('#target');

test(function() {
  assert_equals(container.scrollWidth, target.offsetWidth + 7, "width");
  assert_equals(container.scrollHeight, target.offsetHeight + 9, "height");
}, "overflow, 2D transform");

test(function() {
  target.style.transform = "translateX(17px) translateY(19px)";
  assert_equals(container.scrollWidth, target.offsetWidth + 17, "width");
  assert_equals(container.scrollHeight, target.offsetHeight + 19, "height");
}, "overflow, 2D transform, after css change.");
</script>
